Extract composer submission runner
